requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Markdown==3.2.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.44.0
urllib3==1.25.8
whitenoise==5.0.1
